requirements.txt
google-generativeai
pandas
numpy
python-dotenv
ipykernel
jupyter
